refactor project (#25)
Refactored project to have classes for each LIC to prevent merge conflicts (#25)
