+++ .gitignore
@@ -106,3 +106,4 @@ Temporary Items
 .apdisk
 
 !gradle/wrapper/gradle-wrapper.jar
+.idea/misc.xml
