+++ src/main/java/com/example/utils/MathFunctions.java
@@ -1,0 +1,43 @@
+package com.example.utils;
+import com.example.utils.Point;
+import java.util.List;
+
+public class MathFunctions {
+    public static double calDistance(Point p1, Point p2) {
+        return Math.sqrt(Math.pow(p1.getX() - p2.getX(), 2) + Math.pow(p1.getY() - p2.getY(), 2));
+    }
+
+    /**
+     * Computes the length / size / norm of a vector
+     * @param p vector
+     * @see Point
+     * @return (double) length / size / norm of p
+     */
+    public static double calNorm(Point p) {
+
+        return Math.sqrt(Math.pow(p.getX(), 2) + Math.pow(p.getY(), 2));
+    }
+
+    /**
+     * Checks if a point is inside (or on the border) of a circle centered on (0, 0)
+     * @param p point
+     * @param radius radius of the circle
+     * @see Point
+     * @return (boolean) true if Point p is inside (or on the border) of the circle
+     */
+    public static boolean isInCircle(Point p, double radius) {
+
+        return calNorm(p) <= radius;
+    }
+
+    /**
+     * Checks if three points form a triangle and return the area if it is
+     * @param p1 point
+     * @return (double) area of the triangle if it is a triangle, otherwise return 0
+     */
+    public static double isTriangleArea(Point p1, Point p2, Point p3) {
+        return 0.5 * (p1.getX() * (p2.getY() - p3.getY())
+                + p2.getX() * (p3.getY() - p1.getY())
+                + p3.getX() * (p1.getY() - p2.getY()));
+    }
+}
