+++ src/main/java/com/example/utils/Point.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.utils;
 
 public class Point {
